requirements.txt
torch
transformers 
wandb  
datasets
tqdm
Pillow
huggingface_hub
pandas
torchvision
accelerate
scikit-learn
seaborn
matplotlib